Refactor selectors; divide them two separate files + structure them better per use case
