+++ app/javascript/app/pages/national-context/sectoral-activity/selectors/sectoral-activity-redux-selectors.js
@@ -1,0 +1,6 @@
+export const getEmissionActivities = ({ emissionActivities }) =>
+  emissionActivities && emissionActivities.data;
+
+export const getAdaptation = ({ adaptation }) => adaptation && adaptation.data;
+
+export const getQuery = ({ location }) => location && location.query || null;
